requirements.txt
adal==1.2.7
apize==0.2.7
aspose-3d==23.5.0
azure-core==1.27.1
azure-identity==1.13.0
azure-kusto-data==4.1.4
bpy==3.5.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Cython==0.29.35
dpath==2.1.5
idna==3.4
ijson==3.2.2
importlib-metadata==6.6.0
jaraco.classes==3.2.3
keyring==23.13.1
luau==0.4.21
lxml==4.9.2
midas-data-util==0.3.0
more-itertools==9.1.0
msal==1.22.0
msal-extensions==1.0.0
mutagen==1.46.0
mypy==1.3.0
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.24.3
pandas==2.0.2
pbit==0.2.8
Pillow==9.5.0
portalocker==2.7.0
pyassimp==4.1.4
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.2
roblox-asset-publish==0.2.11
six==1.16.0
tarsafe==0.0.5
toml==0.10.2
tomli==2.0.1
types-PyYAML==6.0.12.9
types-requests==2.31.0.1
types-toml==0.10.8.6
types-urllib3==1.26.25.13
typing_extensions==4.5.0
tzdata==2023.3
unitypackage-extractor==1.1.0
urllib3==1.26.15
zipp==3.15.0
zstandard==0.21.0
